urljoin (complete join), condor disable flag, sqlite timeout, tag developer release

+++ bqcore/bq/util/copylink.py
@@ -26,6 +26,8 @@ def copy_link (*largs):
                 dest = os.path.join (d, os.path.basename(f))
             log.debug ("linking %s to %s", f,dest)
             if os.path.exists(dest):
+                if os.path.abspath(f) == os.path.abspath(dest):
+                    return
                 log.debug ("Found existing file %s: removing .." , dest)
                 os.unlink (dest)
